[add]プライバシーポリシーページの実装およびテスト
