+++ client/cypress/integration/TermsOfUse.spec.ts
@@ -3,7 +3,7 @@ describe('利用規約ページ', () => {
   const terms_of_use_url = `${baseUrl}/help/terms_of_use`;
   it('遷移できること', () => {
     cy.visit(terms_of_use_url);
-    cy.get('[data-testid=terms-of-use-header]');
+    cy.get('[data-testid=header-title]').contains('利用規約').should('exist');
 
     // スナップショットテスト
     cy.matchImageSnapshot('termsOfUse');
