+++ client/src/components/pages/index.ts
@@ -1,0 +1,2 @@
+export { default as PrivacyPolicy } from './helps/PrivacyPolicy';
+export { default as TermsOfUse } from './helps/TermsOfUse';
